Moved each solution into a separate file
Same for tests.
